+++ src/helper/filters/index.ts
@@ -1,0 +1,2 @@
+export { OgsExceptionFilter } from './ogs-exception-filter';
+export { PrismaClientExceptionFilter } from './prisma-client-exception-filter';
